style fixes for demo widget. not perfect, but better. leave buttons for now
